requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
pandas==2.2.0
phonenumbers==8.13.29
PyJWT==2.8.0
python-dotenv==1.0.1
